added changes in handling errors and routes

+++ src/routes/expenses.ts
@@ -1,7 +1,11 @@
 import express, { Response, Request } from "express";
 import { PrismaClient } from "@prisma/client";
 import { PrismaPg } from "@prisma/adapter-pg";
-import { authMiddleware, AuthRequest } from "./auth";
+import {
+  authenticate,
+  AuthenticationRequest,
+} from "../middleware/authentication";
+import { authorize } from "../middleware/authorization";
 
 const connectionString = process.env.DATABASE_URL;
 const router = express.Router();
@@ -26,43 +30,56 @@ const getAllExpenses = async () => {
   return expenses;
 };
 
-router.get("/", authMiddleware, async (_req: AuthRequest, res: Response) => {
-  try {
-    const expenses = await getAllExpenses();
-    res.status(200).send({ response: expenses });
-  } catch (error) {
-    res.status(500).json({ error: error });
+// -----
+
+router.get(
+  "/",
+  authenticate,
+  authorize(["USER", "VISITOR", "ADMIN"]),
+  async (_req: AuthenticationRequest, res: Response) => {
+    try {
+      const expenses = await getAllExpenses();
+      res.status(200).send({ response: expenses });
+    } catch (error) {
+      res.status(500).json({ error: error });
+    }
   }
-});
+);
 
-router.post("/", authMiddleware, async (req: AuthRequest, res: Response) => {
-  const { amount, accountId, categoryId, description } = req.body;
+router.post(
+  "/",
+  authenticate,
+  authorize(["USER", "VISITOR", "ADMIN"]),
+  async (req: AuthenticationRequest, res: Response) => {
+    const { amount, accountId, categoryId, description } = req.body;
 
-  if (!amount || !accountId || !categoryId) {
-    res.status(400).json({ error: "Missing required fields" });
+    if (!amount || !accountId || !categoryId) {
+      res.status(400).json({ error: "Missing required fields" });
+    }
+
+    try {
+      const newExpense = await prisma.expense.create({
+        data: {
+          amount,
+          accountId,
+          categoryId,
+          description,
+          shared: false,
+        },
+      });
+
+      res.status(201).json({ response: newExpense });
+    } catch (error) {
+      res.status(500).json({ error: "Failed to create expense" });
+    }
   }
-
-  try {
-    const newExpense = await prisma.expense.create({
-      data: {
-        amount,
-        accountId,
-        categoryId,
-        description,
-        shared: false,
-      },
-    });
-
-    res.status(201).json({ response: newExpense });
-  } catch (error) {
-    res.status(500).json({ error: "Failed to create expense" });
-  }
-});
+);
 
 router.patch(
   "/:id",
-  authMiddleware,
-  async (req: AuthRequest, res: Response) => {
+  authenticate,
+  authorize(["USER", "VISITOR", "ADMIN"]),
+  async (req: AuthenticationRequest, res: Response) => {
     const { id } = req.params;
     const { amount, accountId, categoryId, description } = req.body;
 
@@ -85,8 +102,9 @@ router.patch(
 
 router.get(
   "/monthly-summary",
-  authMiddleware,
-  async (req: AuthRequest, res: Response) => {
+  authenticate,
+  authorize(["USER", "VISITOR", "ADMIN"]),
+  async (req: AuthenticationRequest, res: Response) => {
     const userId = req.user?.id;
     const { year, month } = req.query;
 
@@ -123,8 +141,9 @@ router.get(
 
 router.get(
   "/months",
-  authMiddleware,
-  async (req: AuthRequest, res: Response) => {
+  authenticate,
+  authorize(["USER", "VISITOR", "ADMIN"]),
+  async (req: AuthenticationRequest, res: Response) => {
     const userId = req.user?.id;
 
     try {
@@ -147,23 +166,28 @@ router.get(
   }
 );
 
-router.delete("/:id", async (req: Request, res: Response) => {
-  const id = req.params.id;
-  if (!id) {
-    res.status(400).send({ error: "Invalid user ID" });
-  }
+router.delete(
+  "/:id",
+  authenticate,
+  authorize(["USER", "VISITOR", "ADMIN"]),
+  async (req: AuthenticationRequest, res: Response) => {
+    const id = req.params.id;
+    if (!id) {
+      res.status(400).send({ error: "Invalid expense ID" });
+    }
 
-  try {
-    await prisma.expense.delete({
-      where: { id },
-    });
+    try {
+      await prisma.expense.delete({
+        where: { id },
+      });
 
-    res
-      .status(200)
-      .json({ message: `Expense with ID ${id} deleted successfully.` });
-  } catch (error) {
-    res.status(404).json({ error: "Expense not found or already deleted." });
+      res
+        .status(200)
+        .json({ message: `Expense with ID ${id} deleted successfully.` });
+    } catch (error) {
+      res.status(404).json({ error: "Expense not found or already deleted." });
+    }
   }
-});
+);
 
 export default router;
